refactor: common domain 패키지 생성  - 도메인 관련 클래스 묶음
