+++ src/main/java/com/prography/lighton/artist/domain/entity/Artist.java
@@ -4,8 +4,8 @@ import com.prography.lighton.artist.domain.entity.enums.ApproveStatus;
 import com.prography.lighton.artist.domain.entity.exception.ArtistRegistrationNotAllowedException;
 import com.prography.lighton.artist.domain.entity.exception.ArtistUpdateNotAllowedException;
 import com.prography.lighton.artist.domain.entity.vo.History;
-import com.prography.lighton.common.BaseEntity;
-import com.prography.lighton.common.vo.RegionInfo;
+import com.prography.lighton.common.domain.BaseEntity;
+import com.prography.lighton.common.domain.vo.RegionInfo;
 import com.prography.lighton.genre.domain.entity.Genre;
 import com.prography.lighton.member.domain.entity.Member;
 import com.prography.lighton.performance.domain.entity.association.PerformanceArtist;
