+++ src/main/java/com/prography/lighton/genre/domain/entity/Genre.java
@@ -1,6 +1,6 @@
 package com.prography.lighton.genre.domain.entity;
 
-import com.prography.lighton.common.BaseEntity;
+import com.prography.lighton.common.domain.BaseEntity;
 import jakarta.persistence.Entity;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
