+++ src/main/java/com/prography/lighton/performance/domain/entity/association/PerformanceArtist.java
@@ -2,7 +2,7 @@ package com.prography.lighton.performance.domain.entity.association;
 
 
 import com.prography.lighton.artist.domain.entity.Artist;
-import com.prography.lighton.common.BaseEntity;
+import com.prography.lighton.common.domain.BaseEntity;
 import com.prography.lighton.performance.domain.entity.Performance;
 import jakarta.persistence.Entity;
 import jakarta.persistence.ManyToOne;
