+++ src/main/java/com/prography/lighton/region/infrastructure/cache/RegionCache.java
@@ -1,6 +1,6 @@
 package com.prography.lighton.region.infrastructure.cache;
 
-import com.prography.lighton.common.vo.RegionInfo;
+import com.prography.lighton.common.domain.vo.RegionInfo;
 import com.prography.lighton.region.domain.repository.SubRegionRepository;
 import com.prography.lighton.region.exception.NoSuchRegionException;
 import com.prography.lighton.region.infrastructure.repository.SubRegionRepository;
